New implementation: 2step, nosplit, critical inputs.

+++ src/vector_simd_vector.hh
@@ -151,7 +151,7 @@ std::ostream& operator<<(std::ostream& os, const vector_simd_vector<T>& v)
 {
   os << "{ ";
   for (size_t i = 0; i < v.size (); ++i)
-    os << v[i] << " ";
+    os << (int) v[i] << " ";
   os << "}";
   return os;
 }
